+++ packages/flexi_cart/test/test.dart
@@ -1,3 +1,4 @@
+import 'flexi_cart_diff_test.dart' as flexi_cart_diff_test;
 import 'flexi_cart_formatter_test.dart' as formatter_test;
 import 'flexi_cart_input_test.dart' as input_test;
 import 'flexi_cart_mixins_test.dart' as mixins_test;
@@ -8,4 +9,5 @@ void main() {
   input_test.main();
   formatter_test.main();
   mixins_test.main();
+  flexi_cart_diff_test.main();
 }